services/notify_svc/pyproject.toml
[project]
name = "notify-svc"
version = "0.1.0"
description = "Notification service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "python-telegram-bot>=20.0",
    "aiosmtplib>=3.0.0",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
